+++ backend/canisters/identity/api/src/updates/create_account_linking_code.rs
@@ -1,0 +1,14 @@
+use crate::account_linking_code::AccountLinkingCode;
+use candid::{CandidType, Deserialize};
+use serde::Serialize;
+use ts_export::ts_export;
+use types::Empty;
+
+pub type Args = Empty;
+
+#[ts_export(identity, create_account_linking_code)]
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub enum Response {
+    Success(AccountLinkingCode),
+    UserNotFound,
+}